move reorganizer to mapper
