+++ mapper/src/main/java/com/ibm/mapper/reorganizer/rules/BlockCipherReorganizer.java
@@ -17,13 +17,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ibm.plugin.translation.reorganizer.rules;
+package com.ibm.mapper.reorganizer.rules;
 
 import com.ibm.mapper.model.Algorithm;
-import com.ibm.mapper.model.AuthenticatedEncryption;
+import com.ibm.mapper.model.BlockCipher;
 import com.ibm.mapper.model.INode;
-import com.ibm.mapper.model.Mac;
-import com.ibm.mapper.model.TagLength;
 import com.ibm.mapper.reorganizer.IReorganizerRule;
 import com.ibm.mapper.reorganizer.builder.ReorganizerRuleBuilder;
 import java.util.ArrayList;
@@ -32,37 +30,34 @@ import java.util.Map;
 import javax.annotation.Nonnull;
 import org.jetbrains.annotations.Unmodifiable;
 
-public final class AeadBlockCipherReorganizer {
+public final class BlockCipherReorganizer {
 
-    private AeadBlockCipherReorganizer() {
+    private BlockCipherReorganizer() {
         // private
     }
 
-    @Nonnull
-    private static final IReorganizerRule MERGE_AE_PARENT_AND_CHILD =
+    private static final IReorganizerRule MERGE_BLOCK_CIPHER_PARENT_AND_CHILD =
             new ReorganizerRuleBuilder()
                     .createReorganizerRule()
-                    .forNodeKind(AuthenticatedEncryption.class)
+                    .forNodeKind(BlockCipher.class)
                     .includingChildren(
                             List.of(
                                     new ReorganizerRuleBuilder()
                                             .createReorganizerRule()
-                                            .forNodeKind(AuthenticatedEncryption.class)
+                                            .forNodeKind(BlockCipher.class)
                                             .noAction()))
                     .perform(
                             (node, parent, roots) -> {
-                                Algorithm newAuthenticatedEncryption =
+                                Algorithm newBlockCipher =
                                         (Algorithm)
                                                 node.getChildren()
-                                                        .get(AuthenticatedEncryption.class)
+                                                        .get(BlockCipher.class)
                                                         .deepCopy();
 
                                 for (Map.Entry<Class<? extends INode>, INode> childKeyValue :
                                         node.getChildren().entrySet()) {
-                                    if (!childKeyValue
-                                            .getKey()
-                                            .equals(AuthenticatedEncryption.class)) {
-                                        newAuthenticatedEncryption.put(childKeyValue.getValue());
+                                    if (!childKeyValue.getKey().equals(BlockCipher.class)) {
+                                        newBlockCipher.put(childKeyValue.getValue());
                                     }
                                 }
 
@@ -72,47 +67,21 @@ public final class AeadBlockCipherReorganizer {
                                     List<INode> rootsCopy = new ArrayList<>(roots);
                                     for (int i = 0; i < rootsCopy.size(); i++) {
                                         if (rootsCopy.get(i).equals(node)) {
-                                            rootsCopy.set(i, newAuthenticatedEncryption);
+                                            rootsCopy.set(i, newBlockCipher);
                                             break;
                                         }
                                     }
                                     return rootsCopy;
                                 } else {
-                                    // Replace the previous AuthenticatedEncryption node
-                                    parent.put(newAuthenticatedEncryption);
+                                    // Replace the previous BlockCipher node
+                                    parent.put(newBlockCipher);
                                     return roots;
                                 }
-                            });
-
-    @Nonnull
-    private static final IReorganizerRule MOVE_TAG_LENGTH_UNDER_MAC =
-            new ReorganizerRuleBuilder()
-                    .createReorganizerRule()
-                    .forNodeKind(AuthenticatedEncryption.class)
-                    .includingChildren(
-                            List.of(
-                                    new ReorganizerRuleBuilder()
-                                            .createReorganizerRule()
-                                            .forNodeKind(Mac.class)
-                                            .noAction(),
-                                    new ReorganizerRuleBuilder()
-                                            .createReorganizerRule()
-                                            .forNodeKind(TagLength.class)
-                                            .noAction()))
-                    .perform(
-                            (node, parent, roots) -> {
-                                TagLength tagLength =
-                                        (TagLength) node.getChildren().get(TagLength.class);
-                                Mac mac = (Mac) node.getChildren().get(Mac.class);
-
-                                mac.put(tagLength);
-                                node.removeChildOfType(TagLength.class);
-                                return roots;
                             });
 
     @Unmodifiable
     @Nonnull
     public static List<IReorganizerRule> rules() {
-        return List.of(MERGE_AE_PARENT_AND_CHILD, MOVE_TAG_LENGTH_UNDER_MAC);
+        return List.of(MERGE_BLOCK_CIPHER_PARENT_AND_CHILD);
     }
 }
