+++ mapper/src/main/java/com/ibm/mapper/reorganizer/rules/MacReorganizer.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ibm.plugin.translation.reorganizer.rules;
+package com.ibm.mapper.reorganizer.rules;
 
 import com.ibm.mapper.model.BlockCipher;
 import com.ibm.mapper.model.BlockSize;
